requirements.txt
uvicorn~=0.23.2
fastapi~=0.103.0
python-dotenv~=1.0.0
pyodbc~=5.2.0